+++ tests/tests_pytorch/strategies/test_fsdp.py
@@ -14,11 +14,7 @@ import torch
 import torch.nn as nn
 from lightning.fabric.plugins.environments import LightningEnvironment
 from lightning.fabric.strategies.fsdp import _is_sharded_checkpoint
-from lightning.fabric.utilities.imports import (
-    _TORCH_GREATER_EQUAL_2_0,
-    _TORCH_GREATER_EQUAL_2_1,
-    _TORCH_GREATER_EQUAL_2_2,
-)
+from lightning.fabric.utilities.imports import _TORCH_GREATER_EQUAL_2_1, _TORCH_GREATER_EQUAL_2_2
 from lightning.fabric.utilities.load import _load_distributed_checkpoint
 from lightning.pytorch import Trainer
 from lightning.pytorch.callbacks import ModelCheckpoint
@@ -28,17 +24,11 @@ from lightning.pytorch.plugins.precision.fsdp import FSDPPrecision
 from lightning.pytorch.strategies import FSDPStrategy
 from lightning.pytorch.trainer.states import TrainerFn
 from lightning.pytorch.utilities.consolidate_checkpoint import _format_checkpoint
-from lightning.pytorch.utilities.exceptions import MisconfigurationException
 from torch.distributed.fsdp.fully_sharded_data_parallel import CPUOffload, FullyShardedDataParallel, MixedPrecision
-from torch.distributed.fsdp.wrap import always_wrap_policy, size_based_auto_wrap_policy, wrap
+from torch.distributed.fsdp.wrap import ModuleWrapPolicy, always_wrap_policy, size_based_auto_wrap_policy, wrap
 from torchmetrics import Accuracy
 
 from tests_pytorch.helpers.runif import RunIf
-
-if _TORCH_GREATER_EQUAL_2_0:
-    from torch.distributed.fsdp.wrap import ModuleWrapPolicy
-else:
-    ModuleWrapPolicy = object
 
 
 class TestFSDPModel(BoringModel):
@@ -88,10 +78,10 @@ class TestFSDPModel(BoringModel):
         assert isinstance(self.trainer.strategy.precision_plugin, FSDPPrecision)
 
         if self.trainer.precision == "16-mixed":
-            param_dtype = None if not _TORCH_GREATER_EQUAL_2_0 else torch.float32
+            param_dtype = torch.float32
             reduce_dtype = buffer_dtype = torch.float16
         elif self.trainer.precision == "bf16-mixed":
-            param_dtype = None if not _TORCH_GREATER_EQUAL_2_0 else torch.float32
+            param_dtype = torch.float32
             reduce_dtype = buffer_dtype = torch.bfloat16
         elif self.trainer.precision == "16-true":
             param_dtype = reduce_dtype = buffer_dtype = torch.float16
@@ -120,10 +110,8 @@ class TestBoringModel(BoringModel):
         self.should_be_wrapped = [wrap_min_params < (32 * 32 + 32), None, wrap_min_params < (32 * 2 + 2)]
 
     def configure_optimizers(self):
-        parameters = self.parameters() if _TORCH_GREATER_EQUAL_2_0 else self.trainer.model.parameters()
-
         # SGD's FSDP optimier state is fixed in https://github.com/pytorch/pytorch/pull/99214
-        return torch.optim.AdamW(parameters, lr=0.1)
+        return torch.optim.AdamW(self.parameters(), lr=0.1)
 
 
 class TestFSDPModelAutoWrapped(TestBoringModel):
@@ -151,10 +139,10 @@ class TestFSDPModelAutoWrapped(TestBoringModel):
         assert isinstance(self.trainer.strategy.precision_plugin, FSDPPrecision)
 
         if self.trainer.precision == "16-mixed":
-            param_dtype = None if not _TORCH_GREATER_EQUAL_2_0 else torch.float32
+            param_dtype = torch.float32
             reduce_dtype = buffer_dtype = torch.float16
         elif self.trainer.precision == "bf16-mixed":
-            param_dtype = None if not _TORCH_GREATER_EQUAL_2_0 else torch.float32
+            param_dtype = torch.float32
             reduce_dtype = buffer_dtype = torch.bfloat16
         elif self.trainer.precision == "16-true":
             param_dtype = reduce_dtype = buffer_dtype = torch.float16
@@ -216,10 +204,7 @@ def _assert_save_equality(trainer, ckpt_path, cls=TestFSDPModel):
 
 def test_invalid_on_cpu(tmp_path, cuda_count_0):
     """Test to ensure that we raise Misconfiguration for FSDP on CPU."""
-    with pytest.raises(
-        MisconfigurationException,
-        match=f"You selected strategy to be `{FSDPStrategy.strategy_name}`, but GPU accelerator is not used.",
-    ):
+    with pytest.raises(ValueError, match="The strategy `fsdp` requires a GPU accelerator"):
         trainer = Trainer(accelerator="cpu", default_root_dir=tmp_path, fast_dev_run=True, strategy="fsdp")
         assert isinstance(trainer.strategy, FSDPStrategy)
         trainer.strategy.setup_environment()
@@ -291,23 +276,12 @@ def test_fsdp_strategy_checkpoint(state_dict_type, precision, tmp_path):
     _run_multiple_stages(trainer, model, os.path.join(tmp_path, "last.ckpt"))
 
 
-if _TORCH_GREATER_EQUAL_2_0:
-
-    def custom_auto_wrap_policy(
-        module,
-        recurse,
-        nonwrapped_numel: int,
-    ) -> bool:
-        return nonwrapped_numel >= 2
-
-else:
-
-    def custom_auto_wrap_policy(
-        module,
-        recurse,
-        unwrapped_params: int,
-    ) -> bool:
-        return unwrapped_params >= 2
+def custom_auto_wrap_policy(
+    module,
+    recurse,
+    nonwrapped_numel: int,
+) -> bool:
+    return nonwrapped_numel >= 2
 
 
 @RunIf(min_cuda_gpus=2, skip_windows=True, standalone=True)
@@ -354,14 +328,6 @@ def test_fsdp_strategy_full_state_dict(tmp_path, wrap_min_params):
             TestFSDPModelAutoWrapped(),
             FSDPStrategy,
             {"auto_wrap_policy": custom_auto_wrap_policy},
-            marks=RunIf(max_torch="2.0.0"),
-            id="autowrap_1x",
-        ),
-        pytest.param(
-            TestFSDPModelAutoWrapped(),
-            FSDPStrategy,
-            {"auto_wrap_policy": custom_auto_wrap_policy},
-            marks=RunIf(min_torch="2.0.0"),
             id="autowrap_2x",
         ),
         pytest.param(
@@ -404,7 +370,7 @@ def test_fsdp_checkpoint_multi_gpus(tmp_path, model, strategy, strategy_cfg):
 @pytest.mark.parametrize("use_orig_params", [None, False, True])
 def test_invalid_parameters_in_optimizer(use_orig_params):
     fsdp_kwargs = {}
-    if _TORCH_GREATER_EQUAL_2_0 and use_orig_params is not None:
+    if use_orig_params is not None:
         fsdp_kwargs = {"use_orig_params": use_orig_params}
 
     trainer = Trainer(
@@ -416,7 +382,7 @@ def test_invalid_parameters_in_optimizer(use_orig_params):
 
     error_context = (
         nullcontext()
-        if _TORCH_GREATER_EQUAL_2_0 and (_TORCH_GREATER_EQUAL_2_1 or use_orig_params is not False)
+        if _TORCH_GREATER_EQUAL_2_1 or use_orig_params is not False
         else pytest.raises(ValueError, match="The optimizer does not seem to reference any FSDP parameters")
     )
 
@@ -435,7 +401,7 @@ def test_invalid_parameters_in_optimizer(use_orig_params):
 
     error_context = (
         nullcontext()
-        if _TORCH_GREATER_EQUAL_2_0 and use_orig_params is not False
+        if use_orig_params is not False
         else pytest.raises(ValueError, match="The optimizer does not seem to reference any FSDP parameters")
     )
 
@@ -534,7 +500,6 @@ def test_fsdp_sharding_strategy():
     assert strategy.sharding_strategy == ShardingStrategy.NO_SHARD
 
 
-@RunIf(min_torch="2.0")
 @pytest.mark.parametrize("sharding_strategy", ["HYBRID_SHARD", "_HYBRID_SHARD_ZERO2"])
 def test_fsdp_hybrid_sharding_strategy(sharding_strategy):
     """Test that the hybrid sharding strategies can only be used with automatic wrapping or a manually specified pg."""
@@ -559,16 +524,11 @@ def test_fsdp_hybrid_sharding_strategy(sharding_strategy):
 
 
 def test_fsdp_use_orig_params():
-    """Test that Lightning enables `use_orig_params` in PyTorch >= 2.0."""
-    with mock.patch("lightning.pytorch.strategies.fsdp._TORCH_GREATER_EQUAL_2_0", False):
-        strategy = FSDPStrategy()
-        assert "use_orig_params" not in strategy.kwargs
-
-    with mock.patch("lightning.pytorch.strategies.fsdp._TORCH_GREATER_EQUAL_2_0", True):
-        strategy = FSDPStrategy()
-        assert strategy.kwargs["use_orig_params"]
-        strategy = FSDPStrategy(use_orig_params=False)
-        assert not strategy.kwargs["use_orig_params"]
+    """Test that Lightning enables `use_orig_params` automatically."""
+    strategy = FSDPStrategy()
+    assert strategy.kwargs["use_orig_params"]
+    strategy = FSDPStrategy(use_orig_params=False)
+    assert not strategy.kwargs["use_orig_params"]
 
 
 @mock.patch("torch.distributed.init_process_group")
@@ -587,7 +547,6 @@ def test_set_timeout(init_process_group_mock):
     )
 
 
-@RunIf(min_torch="2.0")
 @mock.patch("lightning.pytorch.strategies.fsdp._load_raw_module_state")
 def test_fsdp_strategy_load_optimizer_states_multiple(_, tmp_path):
     strategy = FSDPStrategy(parallel_devices=[torch.device("cpu")], state_dict_type="full")
@@ -644,11 +603,8 @@ def test_fsdp_strategy_save_optimizer_states(tmp_path, wrap_min_params):
     if trainer.global_rank != 0:
         assert len(model_state_dict) == 0
 
-    if trainer.global_rank != 0 and _TORCH_GREATER_EQUAL_2_1 or not _TORCH_GREATER_EQUAL_2_0:
+    if trainer.global_rank != 0 and _TORCH_GREATER_EQUAL_2_1:
         assert len(optimizer_state_dict) == 0
-
-    if not _TORCH_GREATER_EQUAL_2_0:
-        return
 
     # restore model to ddp
     model = TestBoringModel()
@@ -674,7 +630,7 @@ def test_fsdp_strategy_save_optimizer_states(tmp_path, wrap_min_params):
 
 @RunIf(min_cuda_gpus=2, skip_windows=True, standalone=True)
 @pytest.mark.parametrize("wrap_min_params", [2, 1024, 100000000])
-def test_fsdp_strategy_load_optimizer_states(tmp_path, wrap_min_params):
+def test_fsdp_strategy_load_optimizer_states(wrap_min_params, tmp_path):
     """Test to ensure that the full state dict and optimizer states can be load when using FSDP strategy.
 
     Based on `wrap_min_params`, the model will be fully wrapped, half wrapped, and not wrapped at all. If the DDP model
@@ -718,10 +674,10 @@ def test_fsdp_strategy_load_optimizer_states(tmp_path, wrap_min_params):
     if trainer.global_rank != 0:
         assert len(restored_model_state_dict) == 0
 
-    if trainer.global_rank != 0 and _TORCH_GREATER_EQUAL_2_1 or not _TORCH_GREATER_EQUAL_2_0:
+    if trainer.global_rank != 0 and _TORCH_GREATER_EQUAL_2_1:
         assert len(restored_optimizer_state_dict) == 0
 
-    if trainer.global_rank == 0 and _TORCH_GREATER_EQUAL_2_0:
+    if trainer.global_rank == 0:
         # assert everything is the same
         assert len(model_state_dict) == len(restored_model_state_dict)
         assert len(optimizer_state_dict) == len(restored_optimizer_state_dict)
@@ -738,14 +694,17 @@ def test_fsdp_strategy_load_optimizer_states(tmp_path, wrap_min_params):
         ("32-true", torch.float32),
     ],
 )
-def test_configure_model(precision, expected_dtype):
+def test_configure_model(precision, expected_dtype, tmp_path):
     """Test that the module under configure_model gets moved to the right device and dtype."""
     trainer = Trainer(
+        default_root_dir=tmp_path,
         accelerator="cuda",
         devices=2,
         strategy=FSDPStrategy(auto_wrap_policy=always_wrap_policy),
         precision=precision,
         max_epochs=1,
+        enable_checkpointing=False,
+        logger=False,
     )
 
     class MyModel(BoringModel):
@@ -770,33 +729,6 @@ def test_configure_model(precision, expected_dtype):
     trainer.fit(model)
 
 
-@mock.patch("lightning.pytorch.strategies.fsdp._TORCH_GREATER_EQUAL_2_0", False)
-@mock.patch("lightning.pytorch.strategies.fsdp.torch.load")
-@mock.patch("lightning.pytorch.strategies.fsdp._load_raw_module_state")
-def test_load_save_optimizer_torch_lt_2_0(_, __, tmp_path):
-    strategy = FSDPStrategy(state_dict_type="full")
-    with pytest.warns(UserWarning, match="does not support saving the optimizer state"):
-        strategy.optimizer_state(Mock())
-
-    file = tmp_path / "test.ckpt"
-    file.touch()
-    trainer = Trainer()
-    trainer.state.fn = TrainerFn.FITTING
-    strategy._lightning_module = Mock(trainer=trainer)
-    with pytest.warns(UserWarning, match="does not support loading the optimizer state"):
-        strategy.load_checkpoint(file)
-
-
-@mock.patch("lightning.pytorch.strategies.fsdp._TORCH_GREATER_EQUAL_2_0", False)
-def test_sharded_state_dict_type_support():
-    """Test that the sharded state dict type is supported."""
-    with pytest.raises(
-        NotImplementedError,
-        match=escape("`FSDPStrategy(state_dict_type='sharded')` is not supported in PyTorch < 2.0"),
-    ):
-        FSDPStrategy(state_dict_type="sharded")
-
-
 def test_save_checkpoint_storage_options(tmp_path):
     """Test that the FSDP strategy does not accept storage options for saving checkpoints."""
     strategy = FSDPStrategy()
@@ -804,7 +736,6 @@ def test_save_checkpoint_storage_options(tmp_path):
         strategy.save_checkpoint(filepath=tmp_path, checkpoint=Mock(), storage_options=Mock())
 
 
-@RunIf(min_torch="2.0.0")
 @mock.patch("lightning.pytorch.strategies.fsdp.FSDPStrategy.broadcast", lambda _, x: x)
 @mock.patch("lightning.pytorch.strategies.fsdp._get_full_state_dict_context")
 @mock.patch("lightning.pytorch.strategies.fsdp._get_sharded_state_dict_context")
@@ -903,7 +834,7 @@ class TestFSDPCheckpointModel(BoringModel):
             torch.testing.assert_close(p0, p1, atol=0, rtol=0, equal_nan=True)
 
 
-@RunIf(min_cuda_gpus=2, standalone=True, min_torch="2.0.0")
+@RunIf(min_cuda_gpus=2, standalone=True)
 def test_save_load_sharded_state_dict(tmp_path):
     """Test FSDP saving and loading with the sharded state dict format."""
     strategy = FSDPStrategy(auto_wrap_policy={nn.Linear}, state_dict_type="sharded")
@@ -959,10 +890,7 @@ def test_fsdp_lazy_load_full_state_dict(_, lazy_load_mock, torch_load_mock, tmp_
     file.touch()
 
     strategy.load_checkpoint(checkpoint_path=file)
-    if _TORCH_GREATER_EQUAL_2_0:
-        lazy_load_mock.assert_called_once()
-    else:
-        torch_load_mock.assert_called_once()
+    lazy_load_mock.assert_called_once()
 
 
 @RunIf(min_cuda_gpus=2, skip_windows=True, standalone=True)
@@ -974,7 +902,7 @@ def test_fsdp_lazy_load_full_state_dict(_, lazy_load_mock, torch_load_mock, tmp_
         pytest.param("bf16-true", torch.bfloat16, marks=RunIf(bf16_cuda=True)),
     ],
 )
-def test_module_init_context(precision, expected_dtype):
+def test_module_init_context(precision, expected_dtype, tmp_path):
     """Test that the module under the init-context gets moved to the right device and dtype."""
 
     class Model(BoringModel):
@@ -990,12 +918,15 @@ def test_module_init_context(precision, expected_dtype):
 
     def _run_setup_assertions(empty_init, expected_device):
         trainer = Trainer(
+            default_root_dir=tmp_path,
             accelerator="cuda",
             devices=2,
             strategy=FSDPStrategy(auto_wrap_policy={torch.nn.Linear}),
             precision=precision,
             max_steps=1,
             barebones=True,
+            enable_checkpointing=False,
+            logger=False,
         )
         with trainer.init_module(empty_init=empty_init):
             model = Model()